TF2/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
cloudmesh-common==4.3.91
colorama==0.4.4
dill==0.3.4
entrypoints==0.3
flatbuffers==2.0
fonttools==4.25.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
humanize==3.14.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
oyaml==1.0
pandas==1.3.5
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
PyQt5==5.15.2
PyQt5_sip==4.19.25
PyQtChart==5.15.2
python-hostlist==1.21
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.1
simplejson==3.17.6
six==1.15.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
tensorflow-metadata==1.6.0
tensorflow-metal==0.3.0
termcolor==1.1.0
terminado==0.9.4
tqdm==4.62.3
typeguard==2.13.3
typing_extensions==4.0.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.13.3
